Refactor: added an api struct for db
In the attempt to add mock DB, modified the code for future to be able to support both mock DB and real DB via a common object. This object could have various functions which connects to database. In my code, the Add Product handler and models.go needed this, to communicate to mock DB. But this was unsuccessful attempt to be able to mimic the DB with the Dataa Dog Mocking Library

Signed-off-by: Utkarsh Chourasia <5.utkarshchourasia@gmail.com>